Commented out check for mismatched sender MAC address in ARP message body and Ethernet header
This functionality is similar to Cisco's ip arp inspection validate src-mac feature, which is an optional DAI feature. The check can be uncommented if administrators wish to enable it for ARP spoofing prevention, but it may interfere with DHCP snooping in bridged environments. This code remains to give users the flexibility in enabling or disabling it as needed.

Signed-off-by: Korel <Korelucpinar@gmail.com>