DAI Feature: Implement trusted interface list management
Added functionality that allows users to manage a list of trusted network interfaces. Implemented insert_trusted_interface() to add interfaces to a trusted list and added find_trusted_interface() to check if an interface exists in the list. This feature is intended to be combined with DAI rate limiting in future commits.

Signed-off-by: Korel <Korelucpinar@gmail.com>