Generalize ARP Hook to Bridge Hook for Trusted and Untrusted Interfaces
Replaced the ARP hook with a more general Bridge hook to support the expansion of Trusted and Untrusted Interfaces. The Bridge hook maintains all the functionality of the ARP hook while incorporating additional checks. It first checks the trust state of the interface where incoming packets are received. If the interface is trusted, all following DAI checks are skipped. If the interface is untrusted, packet acceptance is determined by the interface's rate limit status (Rate Limiting is to be added in a future commit). If the interface hasn't been rate-limited then confirm the validity of the packet using the DHCP or Static Entry tables.

Signed-off-by: Korel <Korelucpinar@gmail.com>